requirements_web.txt
redis==2.10.5
rq==0.6.0
djangorestframework==3.4
